Fix: Auto login was working only after going to Auth screen
